+++ packages/cluster/src/commands/compose.js
@@ -1,6 +1,6 @@
-const chalk = require('chalk')
 const path = require('path')
-const { reporter, exec, tryCatchAsync } = require('@dhis2/cli-helpers-engine')
+const { spawn } = require('child_process')
+const { reporter, tryCatchAsync } = require('@dhis2/cli-helpers-engine')
 const {
     initDockerComposeCache,
     makeComposeProject,
@@ -8,11 +8,12 @@ const {
     resolveConfiguration,
 } = require('../common')
 
-const defaults = require('../defaults')
-
 const run = async function(argv) {
-    const { name, service } = argv
+    const { name, _ } = argv
     const cfg = await resolveConfiguration(argv)
+
+    const args = _.slice(_.findIndex(x => x === 'compose') + 1)
+    reporter.debug('Passing arguments to docker-compose', args)
 
     const cacheLocation = await initDockerComposeCache({
         composeProjectName: name,
@@ -25,38 +26,36 @@ const run = async function(argv) {
         reporter.error('Failed to initialize cache...')
         process.exit(1)
     }
-    reporter.info(`Spinning up cluster version ${chalk.cyan(name)}`)
     const res = await tryCatchAsync(
         'exec(docker-compose)',
-        exec({
-            cmd: 'docker-compose',
-            args: [
+        spawn(
+            'docker-compose',
+            [
                 '-p',
                 makeComposeProject(name),
                 '-f',
                 path.join(cacheLocation, 'docker-compose.yml'),
-                'restart',
-            ].concat(service ? [service] : []),
-            env: makeEnvironment(cfg),
-            pipe: true,
-        })
+                ...args,
+            ],
+            {
+                env: {
+                    ...process.env,
+                    ...makeEnvironment(cfg),
+                },
+                stdio: 'inherit',
+            }
+        )
     )
     if (res.err) {
-        reporter.error('Failed to restart cluster service(s)')
+        reporter.error('Failed to run docker-compose command')
         process.exit(1)
     }
 }
 
 module.exports = {
-    command: 'restart <name> [service]',
-    desc: 'Restart a cluster or cluster service',
-    aliases: 'r',
-    builder: {
-        port: {
-            alias: 'p',
-            desc: `Specify the port on which to expose the DHIS2 instance (default: ${defaults.port})`,
-            type: 'integer',
-        },
-    },
+    command: 'compose <name>',
+    desc:
+        'Run arbitrary docker-compose commands against a DHIS2 cluster.\nNOTE: pass -- after <name>',
+    aliases: 'c',
     handler: run,
 }