+++ packages/cluster/src/commands/logs.js
@@ -4,26 +4,33 @@ const { reporter, exec, tryCatchAsync } = require('@dhis2/cli-helpers-engine')
 const {
     initDockerComposeCache,
     makeComposeProject,
+    makeEnvironment,
     resolveConfiguration,
 } = require('../common')
-const defaults = require('../defaults')
 
-const run = async function({ service, name, cluster, ...argv }) {
-    const { dockerComposeRepository } = resolveConfiguration(argv, {}, cluster)
+const run = async function(argv) {
+    const { service, name } = argv
+    const cfg = await resolveConfiguration(argv)
+
     const cacheLocation = await initDockerComposeCache({
+        composeProjectName: name,
         cache: argv.getCache(),
-        dockerComposeRepository,
+        dockerComposeRepository: cfg.dockerComposeRepository,
+        dockerComposeDirectory: cfg.dockerComposeDirectory,
         force: false,
     })
+
     if (!cacheLocation) {
         reporter.error('Failed to initialize cache...')
         process.exit(1)
     }
+
     reporter.info(
         `Reading logs from cluster version ${chalk.cyan(name)}${
             service ? ` <${service}>` : ''
         }`
     )
+
     const res = await tryCatchAsync(
         'exec(docker-compose)',
         exec({
@@ -36,10 +43,7 @@ const run = async function({ service, name, cluster, ...argv }) {
                 'logs',
                 '-f',
             ].concat(service ? [service] : []),
-            env: {
-                DHIS2_CORE_NAME: name,
-                DHIS2_CORE_PORT: 8000, // doesn't matter
-            },
+            env: makeEnvironment(cfg),
             pipe: true,
         })
     )
@@ -52,6 +56,6 @@ const run = async function({ service, name, cluster, ...argv }) {
 module.exports = {
     command: 'logs <name> [service]',
     desc: 'Tail the logs from a given service',
-    aliases: 'r',
+    aliases: 'l',
     handler: run,
 }