+++ packages/cluster/src/defaults.js
@@ -7,8 +7,9 @@ module.exports = {
     customContext: false,
     update: false,
     seed: false,
+    dockerComposeDirectory: 'cluster',
     dockerComposeRepository:
-        'https://github.com/amcgee/dhis2-backend/archive/master.tar.gz',
+        'https://github.com/dhis2/docker-compose/archive/master.tar.gz',
     demoDatabaseURL:
         'https://github.com/dhis2/dhis2-demo-db/blob/master/sierra-leone/{version}/dhis2-db-sierra-leone.sql.gz?raw=true',
 }