+++ packages/cluster/src/helpers/db/index.js
@@ -1,0 +1,2 @@
+module.exports.backup = require('./backup')
+module.exports.restore = require('./restore')